Exercise 10 (page 80)

+++ chapter 3/exercise 9.cpp
@@ -4,9 +4,9 @@
 
 int main()
 {
-    uint64_t last = 2;
+    double last = 2;
 
-    for (uint64_t i = 0; i < 64; i++)
+    for (double i = 0; i < 64; i++)
     {
         cout << i + 1 << ' ' << last << '\n';
         last *= 2;
@@ -14,6 +14,4 @@ int main()
     
 }
 
-// For 1.000         - 10
-// For 1.000.000     - 20
-// For 1.000.000.000 - 27
+// The maximun numebr you can count with the double type is infinite
